% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBatchSd.R
\name{getBatchSd}
\alias{getBatchSd}
\title{getBatchSd}
\usage{
getBatchSd(x, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}

\item{list}{of dataframes}
}
\value{
dataframe
}
\description{
takes a list of dataframes containing batch information and time
and returns one dataframe with the standard deviations
of these data, sorted by hour
}
\examples{
getBatchSd(a_wo_outlier)
}
