% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fermdata.R
\name{plot_fermdata}
\alias{plot_fermdata}
\title{plot_fermdata}
\usage{
plot_fermdata(x, cumulate = "co2", type = NA, filename = NA,
  width = 480, height = 480, res = 72)
}
\arguments{
\item{x}{a list of dataframes}

\item{cumulate}{a name of a column}

\item{type}{the type of saved format}

\item{filename}{the filename}

\item{width}{picture width}

\item{height}{picture height}

\item{res}{picture resolution}
}
\value{
plot
}
\description{
Takes a list of Dataframes, containing Batch-relevant information, which substance
gets cumulated (co2 per default) and if type and filename are given, saves the results
as a png or jpeg.
}
\examples{
plot_fermdata(a_wo_outlier, cumulate="co2", type = NA, filename = NA)
}
