% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCompoundData.R
\name{readCompoundData}
\alias{readCompoundData}
\title{readCompoundData}
\usage{
readCompoundData(x)
}
\arguments{
\item{a}{file}
}
\value{
x a dataframe
}
\description{
reads a file containing the compound data
}
\examples{
readCompoundData("inst/extdata/A/Compounds.txt")
}
