% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCarbonBalance.R
\name{saveCarbonBalance}
\alias{saveCarbonBalance}
\title{saveCarbonBalance}
\usage{
saveCarbonBalance(x, y)
}
\arguments{
\item{a}{dataframe}

\item{a}{path}
}
\value{
a .csv
}
\description{
Function to save a dataframe as a csv, sep="," and dec="."
}
\examples{
saveCarbonBalance(a_carbon_balance, paste(output_dir, "a_carbon_balance.csv", sep = ""))
}
