% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTestSd.R
\name{getTestSd}
\alias{getTestSd}
\title{getTestSd}
\usage{
getTestSd(x, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}

\item{a}{list of dataframes}
}
\value{
a dataframe
}
\description{
Auxiliary Function for getBatchSd.
}
\examples{
getTestSd(a_wo_outlier)
}
