% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fermdata.R
\name{plot_fermdata}
\alias{plot_fermdata}
\title{plot_fermdata}
\usage{
plot_fermdata(x, cumulate = "co2", type = NA, filename = NA,
  width = 480, height = 480, res = 72)
}
\arguments{
\item{res}{}

\item{list}{of dataframes}
}
\value{
plot
}
\description{
Takes a list of Dataframes, containing Batch-relevant information, which substance
gets cumulated (co2 per default) and if type and filename are given, saves the results
as a png or jpeg.
}
\examples{
plot_fermdata(a_wo_outlier, cumulate="co2", type = "png", filename = paste(output_dir, "a.png", sep = ""), width = 1200,height = 800)
}
